+++ packages/calcite-components/src/components/date-picker-month-header/date-picker-month-header.tsx
@@ -1,3 +1,4 @@
+// @ts-strict-ignore
 import {
   calciteSpacingBase,
   calciteSpacingXxs,