+++ packages/calcite-components/src/components/flow/flow.tsx
@@ -1,3 +1,4 @@
+// @ts-strict-ignore
 import { PropertyValues } from "lit";
 import { LitElement, property, h, method, state, JsxNode } from "@arcgis/lumina";
 import { createObserver } from "../../utils/observers";
@@ -102,9 +103,9 @@ export class Flow extends LitElement implements LoadableComponent {
     await componentFocusable(this);
 
     const { items } = this;
-    const activeItem = items[items.length - 1];
+    const selectedItem = items[this.selectedIndex];
 
-    return activeItem?.setFocus();
+    return selectedItem?.setFocus();
   }
 
   // #endregion