+++ packages/calcite-components/src/components/tab-nav/tab-nav.tsx
@@ -1,3 +1,4 @@
+// @ts-strict-ignore
 import {
   calciteSize24,
   calciteSize32,