+++ packages/calcite-components/src/custom-theme.stories.ts
@@ -14,7 +14,9 @@ import { alertTokens, alert } from "./custom-theme/alert";
 import { accordionItemTokens } from "./custom-theme/accordion-item";
 import { accordion, accordionTokens } from "./custom-theme/accordion";
 import { autocomplete, autocompleteTokens } from "./custom-theme/autocomplete";
+import { block, blockTokens } from "./custom-theme/block";
 import { buttons } from "./custom-theme/button";
+import { blockSection, blockSectionTokens } from "./custom-theme/block-section";
 import { calciteSwitch } from "./custom-theme/switch";
 import { card, cardThumbnail, cardTokens } from "./custom-theme/card";
 import { checkbox, checkboxTokens } from "./custom-theme/checkbox";
@@ -31,7 +33,7 @@ import { label, labelTokens } from "./custom-theme/label";
 import { link, linkTokens } from "./custom-theme/link";
 import { list, listTokens } from "./custom-theme/list";
 import { loader, loaderTokens } from "./custom-theme/loader";
-import { notices } from "./custom-theme/notice";
+import { notice, noticeTokens } from "./custom-theme/notice";
 import { pagination } from "./custom-theme/pagination";
 import { popover, popoverTokens } from "./custom-theme/popover";
 import { progress, progressTokens } from "./custom-theme/progress";
@@ -111,7 +113,7 @@ const kitchenSink = (args: Record<string, string>, useTestValues = false) =>
     </style>
     <div class="demo">
       <div class="demo-column">
-        ${accordion} ${actionBar} ${autocomplete} ${notices} ${segmentedControl}
+        ${accordion} ${actionBar} ${autocomplete} ${notice} ${segmentedControl}
         <div style="display: flex">
           ${actionPad}
           <div style="width: 40px; height: 40px;">${actionMenu}</div>
@@ -131,8 +133,8 @@ const kitchenSink = (args: Record<string, string>, useTestValues = false) =>
         ${avatarInitials} ${avatarThumbnail} ${progress} ${handle} ${textArea} ${popover} ${tile} ${tooltip}
         ${comboboxItem}
       </div>
-      <div class="demo-column">${navigation} ${navigationLogos} ${navigationUsers}</div>
-      ${alert}
+      <div class="demo-column">${navigation} ${navigationLogos} ${navigationUsers} ${blockSection} ${block}</div>
+      <div class="demo-column"><div class="demo-column">${alert}</div></div>
     </div>
   </div>`;
 
@@ -148,6 +150,8 @@ const componentTokens = {
   ...alertTokens,
   ...autocompleteTokens,
   ...avatarTokens,
+  ...blockSectionTokens,
+  ...blockTokens,
   ...cardTokens,
   ...checkboxTokens,
   ...chipTokens,
@@ -160,6 +164,7 @@ const componentTokens = {
   ...navigationLogoTokens,
   ...navigationTokens,
   ...navigationUserTokens,
+  ...noticeTokens,
   ...popoverTokens,
   ...progressTokens,
   ...sliderTokens,