+++ packages/calcite-design-tokens/support/token-transformer/styleDictionary/formatter/docs.ts
@@ -1,3 +1,4 @@
+// @ts-strict-ignore
 import { Core as StyleDictionary } from "style-dictionary";
 import prettierSync from "@prettier/sync";